% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugAdherence-package.R, R/DrugAdherence.R
\docType{package}
\name{DrugAdherence-package}
\alias{DrugAdherence-package}
\title{DrugAdherence: DrugAdherence: To learn patterns of Medication Compliance, Aderence and Persistence}
\description{
The DrugAdherence R package analyzes medication compliance and persistence in cohorts. It assesses drug utilization from first exposure, calculates key metrics like medication possession ratio and proportional days covered, and enhances insights into real-world adherence to improve patient outcomes.

The DrugAdherence R package analyzes medication compliance and persistence in cohorts. It assesses drug utilization from first exposure, calculates key metrics like medication possession ratio and proportional days covered, and enhances insights into real-world adherence to improve patient outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DrugAdherence/}
  \item \url{https://github.com/OHDSI/DrugAdherence}
  \item Report bugs at \url{https://github.com/OHDSI/DrugAdherence/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DrugAdherence/}
  \item \url{https://github.com/OHDSI/DrugAdherence}
  \item Report bugs at \url{https://github.com/OHDSI/DrugAdherence/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
